Add feature detection to OxiPNG
