+++ codecs/oxipng/build.sh
@@ -3,7 +3,7 @@
 set -e
 
 rm -rf pkg,{-parallel}
-wasm-pack build -t web
+wasm-pack build
 RUSTFLAGS='-C target-feature=+atomics,+bulk-memory' wasm-pack build -t web -d pkg-parallel -- -Z build-std=panic_abort,std --features=parallel
 sed -i "s|input = import.meta.url.replace(/\\\.js$/, '_bg.wasm');||" pkg{,-parallel}/squoosh_oxipng.js
 rm pkg{,-parallel}/.gitignore
