+++ codecs/oxipng/pkg/squoosh_oxipng.d.ts
@@ -6,25 +6,3 @@
 * @returns {Uint8Array}
 */
 export function optimise(data: Uint8Array, level: number): Uint8Array;
-
-export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembly.Module;
-
-export interface InitOutput {
-  readonly memory: WebAssembly.Memory;
-  readonly optimise: (a: number, b: number, c: number, d: number) => void;
-  readonly malloc: (a: number) => number;
-  readonly free: (a: number) => void;
-  readonly __wbindgen_malloc: (a: number) => number;
-  readonly __wbindgen_free: (a: number, b: number) => void;
-}
-
-/**
-* If `module_or_path` is {RequestInfo} or {URL}, makes a request and
-* for everything else, calls `WebAssembly.instantiate` directly.
-*
-* @param {InitInput | Promise<InitInput>} module_or_path
-*
-* @returns {Promise<InitOutput>}
-*/
-export default function init (module_or_path?: InitInput | Promise<InitInput>): Promise<InitOutput>;
-
